call to get FOrmats and Types list in MAin Activity

+++ app/src/main/java/lt/pusnis/multibarcodereader/model/MbrFormats.java
@@ -15,7 +15,7 @@ public class MbrFormats {
         return id;
     }
 
-    public void setId(long id) {
+    public void setId(int id) {
         this.id = id;
     }
